[inductor] Install guards on both cases of View.handle_negative_index
This branch is not an optimization, it's a correctness issue so there should be
a guards installed for both sides of the branch. Otherwise we could have an
expression like `(s0 - s1)` that is initially positive, then becomes negative
with a new set of shapes and now references an invalid index.

ghstack-source-id: 6b06f6745f50638e8a57c427fc13ede590b3b96d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103780